third_party/blink/web_tests/animations/interpolation/list-style-image-interpolation-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Transitions: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.3) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.3 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
[FAIL] CSS Transitions: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.5) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.5 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
[FAIL] CSS Transitions: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.6) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.6 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
[FAIL] CSS Transitions with transition: all: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.3) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.3 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
[FAIL] CSS Transitions with transition: all: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.5) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.5 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
[FAIL] CSS Transitions with transition: all: property <list-style-image> from [url(../resources/stripes-20.png)] to [url(../resources/stripes-20.png)] at (0.6) is [url(file:///.../stripes-20.png)]
  assert_equals: expected "- webkit - cross - fade ( url ( file : / / / ... / stripes - 20.png ) , url ( file : / / / ... / stripes - 20.png ) , 0.6 ) " but got "url ( file : / / / ... / stripes - 20.png ) "
Harness: the test ran to completion.
